added sessions controller, middlewares and user service
